requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
celery==5.3.4
django-celery-beat==2.5.0

